fix(android): use float division for download progress ratio
 was always receiving 0 for in-flight
downloads because the ratio was computed with long/long integer
division, which truncates to 0 until the transfer completes.

The count-based throttle in ReactNativeBlobUtilProgressConfig is gated
on , so that guard never became true and
 stayed at its default of . The result is that
 was silently ignored on Android and every
consumer fell back to interval-only reporting.

Cast to float at the three sites that compute the ratio:

- ReactNativeBlobUtilFileResp: download-to-file responses
- ReactNativeBlobUtilDefaultResp: in-memory responses
- ReactNativeBlobUtilReq: DownloadManager polling loop

DefaultResp additionally guarded against chunked responses, where
contentLength() is -1 and the old expression produced a negative
ratio. Both the old and new code take the interval-only path there, so
this is a clarification rather than a behaviour change.

Default usage (, count -1) is unaffected.

+++ android/src/main/java/com/ReactNativeBlobUtil/ReactNativeBlobUtilReq.java
@@ -217,7 +217,7 @@ public class ReactNativeBlobUtilReq extends BroadcastReceiver implements Runnabl
                             cursor.close();
 
                             ReactNativeBlobUtilProgressConfig reportConfig = getReportProgress(taskId);
-                            float progress = (total > 0) ? written / total : 0;
+                            float progress = (total > 0) ? (float) written / total : 0;
 
                             if (reportConfig != null && reportConfig.shouldReport(progress /* progress */)) {
                                 WritableMap args = Arguments.createMap();
